% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiRSDA.R
\name{getVHP}
\alias{getVHP}
\title{title}
\usage{
getVHP(y, buffer = NULL, dates = NULL, path = NULL, returnRast = TRUE,
  ...)
}
\arguments{
\item{y}{points represented by a two-column matrix or data.frame, or `SpatialPoints`; `SpatialPolygons`; `SpatialLines`; `Extent`; or a numeric vector representing cell numbers}

\item{path}{the path leading to the `VHP.G16.C07.SM.csv` file. If `NULL` (default), the funcition will search in default R drive folder.}
}
\value{
...
}
\description{
Title
}
\details{
...
}
